fnished sgr and changes to bags order in product list
